requirements.txt
python-telegram-bot==20.7
mysql-connector-python==8.1.0
apscheduler==3.10.4